Change log level in read fail message in dmic_nrfx_pdm.c
The log level used to indicate no data was read is LOG_ERR, but really this is not an error if your application does not want to block (i.e., is doing other things in the background). Changing it to LOG_DBG I believe is more fitting for that use case and doesn't flood the log with error messages.